refactoring: PopupWithImage implemented, old code removed
