+++ src/components/index.js
@@ -1,4 +1,3 @@
-//new code
 import '../pages/index.css';
 import {
   apiConfig,
@@ -15,12 +14,13 @@ import {
   profilePictureButton
 } from './utils/constants.js';
 
-import Api from './Api_class.js'; //TODO: fix path
+import Api from './Api.js';
 import UserInfo from './UserInfo.js';
-import Card from './Card_class.js';
+import Card from './Card.js';
 import FormValidator from './FormValidator';
 import Section from './Section.js';
 import PopupWithForm from './PopupWithForm.js';
+import PopupWithImage from './PopupWithImage';
 
 
 
@@ -33,6 +33,10 @@ const userInfo = new UserInfo({nameSelector: profileSelectors.name,
   avatarSelector: profileSelectors.pic
 
 });
+
+const popupWithImage = new PopupWithImage('.popup_pic');
+popupWithImage.setEventListeners();
+
 
 api.getProfileData()
 .then((data) => {
@@ -70,6 +74,9 @@ api.getProfileData()
                     .catch((err) => {
                       console.log(err);
                     });
+                },
+                handlePopup: () => {
+                  popupWithImage.open(item.name, item.link);
                 }
               });
               return card.renderCard();
@@ -186,6 +193,9 @@ const cardPopup = new PopupWithForm('.popup_add', (inputs) => {
                     .catch((err) => {
                       console.log(err);
                     });
+                },
+                handlePopup: () => {
+                  popupWithImage.open(item.name, item.link);
                 }
               });
               return card.renderCard();
@@ -212,152 +222,3 @@ addButton.addEventListener('click', () =>{
   cardPopupValidator.enableValidation();
   cardPopupValidator.disableButton();
 });
-
-//old code
-
-// import {openPopup, closePopup} from './modal.js';
-// import {renderCard, addCardOnPage} from './cards.js';
-// //import {enableValidation, disableButton} from './validate.js';
-// import {getInitialCards, getProfileData, editProfile, publishCard, editAvatar} from './api.js'
-
-
-
-// const popupCloseButtons = document.querySelectorAll('.popup__close-button');
-
-// const popupAdd = document.querySelector('.popup_add');
-// const addForm = popupAdd.querySelector('.form');
-// const addFormButton = addForm.querySelector('.form__submit-button');
-
-// const popupProfile = document.querySelector('.popup_profile');
-// const profileFormButton = profileForm.querySelector('.form__submit-button');
-// const profileFormName = profileForm.querySelector('.form__input-text_field_name');
-// const profileFormCaption = profileForm.querySelector('.form__input-text_field_caption');
-// const profileName = document.querySelector('.profile__name');
-// const profileCaption = document.querySelector('.profile__caption');
-// const profilePic = document.querySelector('.profile__picture');
-// const pictureFormCaption = addForm.querySelector('.form__input-text_field_caption');
-// const pictureFormLink = addForm.querySelector('.form__input-text_field_link');
-
-// const profilePictureFormButton = profilePictureForm.querySelector('.form__submit-button');
-// const profilePictureInput = profilePictureForm.querySelector('.form__input-text_field_link');
-
-// const profileData = {};
-
-
-
-// function handleEditProfilePopup () {
-//   openPopup(popupProfile);
-//   profileFormName.value = profileData.name;
-//   profileFormCaption.value = profileData.about;
-// }
-
-// function handleEditProfilePicturePopup () {
-//   openPopup(popupProfilePic);
-// }
-
-// function handleAddPopup () {
-//   openPopup(popupAdd);
-// }
-
-// function submitProfileForm (evt) {
-//   changeSubmitButtonText(profileFormButton, 'Сохранение...');
-//   evt.preventDefault();
-//   editProfile(profileFormName.value, profileFormCaption.value)
-//   .then((data) => {
-//     setProfileData(data);
-//     renderProfile(data);
-//     closePopup(popupProfile);
-//     disableButton(profileFormButton, validationElements.inactiveButtonClass);
-//   })
-//   .catch((err) => {
-//     console.log(err);
-//   })
-//   .finally(function () {
-//     changeSubmitButtonText(profileFormButton, 'Сохранить');
-//   });
-// }
-
-// function addNewCard (evt) {
-//   changeSubmitButtonText(addFormButton, 'Сохранение...');
-//   evt.preventDefault();
-//   publishCard(pictureFormCaption.value, pictureFormLink.value)
-//   .then((data) => {
-//     addCardOnPage(renderCard(data, profileData.id))
-//     closePopup(popupAdd);
-//     addForm.reset();
-//     disableButton(addFormButton, validationElements.inactiveButtonClass);
-//   })
-//   .catch((err) => {
-//     console.log(err);
-//   })
-//   .finally(function () {
-//     changeSubmitButtonText(addFormButton, 'Сохранить');
-//   });
-// }
-
-// function submitProfilePictureForm (evt) {
-//   changeSubmitButtonText(profilePictureFormButton, 'Сохранение...');
-//   evt.preventDefault();
-//   editAvatar(profilePictureInput.value)
-//   .then((data) => {
-//     setProfileData(data);
-//     renderProfile(data);
-//     closePopup(popupProfilePic);
-//     disableButton(profilePictureFormButton, validationElements.inactiveButtonClass);
-//   })
-//   .catch((err) => {
-//     console.log(err);
-//   })
-//   .finally(function () {
-//     changeSubmitButtonText(profilePictureFormButton, 'Сохранить');
-//   });
-// }
-
-
-// addButton.addEventListener('click', handleAddPopup);
-// profilePictureButton.addEventListener('click', handleEditProfilePicturePopup);
-
-// popupCloseButtons.forEach(closeButton => {
-//   const popup = closeButton.closest('.popup');
-//   closeButton.addEventListener('click',() => closePopup(popup));
-// });
-
-//profileForm.addEventListener('submit', submitProfileForm);
-// addForm.addEventListener('submit', addNewCard);
-// profilePictureForm.addEventListener('submit', submitProfilePictureForm);
-
-//enableValidation(validationElements);
-
-// getProfileData()
-// .then((data) => {
-//   //setProfileData(data);
-//   //renderProfile(profileData);
-//   getInitialCards()
-//   .then((data) => {
-//     data.reverse().forEach(item => addCardOnPage(renderCard(item, profileData.id)));
-//     })
-//     .catch((err) => {
-//       console.log(err);
-//     });
-// })
-//   .catch((err) => {
-//     console.log(err);
-// });
-
-// function renderProfile (profileData) {
-//   profileName.textContent = profileData.name;
-//   profileCaption.textContent = profileData.about;
-//   profilePic.src = profileData.avatar;
-// }
-
-// function setProfileData (data) {
-//   profileData.name = data.name;
-//   profileData.about = data.about;
-//   profileData.avatar = data.avatar;
-//   profileData.id = data._id;
-// }
-
-// function changeSubmitButtonText (buttonElement, text) {
-//   buttonElement.value = text;
-// }
-
